fixed that shit

+++ comment_member_delete.php
@@ -1,6 +1,10 @@
 <?php
+if(getUserIdFromSession() == true) {
+  deleteComment($commentId);
+  header('Location: index.php?function=entries_member_details&bid='.$blogId.'&eid='.$entryId.'');
+} else {
+  die('Bitte zuerst <a href="index.php?function=login">einloggen</a>');
+}
 
-deleteComment($commentId);
-header('Location: index.php?function=entries_member_details&bid='.$blogId.'&eid='.$entryId.'');
 
   ?>
