+++ entries_member_delete.php
@@ -1,6 +1,8 @@
 <?php
-
-deleteEntry($entryId);
-header('Location: index.php?function=entries_member&bid='.$_SESSION['uid']);
-
+if(getUserIdFromSession() == true) {
+  deleteEntry($entryId);
+  header('Location: index.php?function=entries_member&bid='.$_SESSION['uid']);
+} else {
+  die('Bitte zuerst <a href="index.php?function=login">einloggen</a>');
+}
   ?>
